Add payment methods to request model
