+++ src/Vendr.PaymentProviders.Dibs/Easy/Api/Models/DibsPaymentRequest.cs
@@ -1,5 +1,4 @@
 ﻿using Newtonsoft.Json;
-using System.Collections.Generic;
 
 namespace Vendr.PaymentProviders.Dibs.Easy.Api.Models
 {
@@ -16,48 +15,9 @@ namespace Vendr.PaymentProviders.Dibs.Easy.Api.Models
 
         [JsonProperty("notifications")]
         public DibsNotifications Notifications { get; set; }
-    }
 
-    public class DibsOrder
-    {
-        [JsonProperty("reference")]
-        public string Reference { get; set; }
-
-        [JsonProperty("currency")]
-        public string Currency { get; set; }
-
-        [JsonProperty("amount")]
-        public int Amount { get; set; }
-
-        [JsonProperty("items")]
-        public DibsOrderItem[] Items { get; set; }
-    }
-
-    public class DibsOrderItem
-    {
-        [JsonProperty("reference")]
-        public string Reference { get; set; }
-
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("quantity")]
-        public int Quantity { get; set; }
-
-        [JsonProperty("unit")]
-        public string Unit { get; set; }
-
-        [JsonProperty("unitPrice")]
-        public int UnitPrice { get; set; }
-
-        [JsonProperty("taxRate")]
-        public int TaxRate { get; set; }
-
-        [JsonProperty("grossTotalAmount")]
-        public int GrossTotalAmount { get; set; }
-
-        [JsonProperty("netTotalAmount")]
-        public int NetTotalAmount { get; set; }
+        [JsonProperty("paymentMethods")]
+        public DibsPaymentMethod[] PaymentMethods { get; set; }
     }
 
     public class DibsConsumer
